[Bifrost] Move TailState to restate-types
This allows TailState to be used in log-server as well
