+++ crates/bifrost/src/types.rs
@@ -78,81 +78,6 @@ pub struct FindTailAttributes {
     // TODO: consistent_read: bool,
 }
 
-/// Represents the state of the tail of the loglet.
-#[derive(Clone, Debug)]
-pub enum TailState<Offset = Lsn> {
-    /// Loglet is open for appends
-    Open(Offset),
-    /// Loglet is sealed. This offset if the durable tail.
-    Sealed(Offset),
-}
-
-impl<Offset: SequenceNumber> TailState<Offset> {
-    pub fn new(sealed: bool, offset: Offset) -> Self {
-        if sealed {
-            TailState::Sealed(offset)
-        } else {
-            TailState::Open(offset)
-        }
-    }
-
-    /// Combines two TailStates together
-    ///
-    /// Only applies updates to the value according to the following rules:
-    ///   - Offsets can only move forward.
-    ///   - Tail cannot be unsealed once sealed.
-    ///
-    /// Returns true if the state was updated
-    pub fn combine(&mut self, sealed: bool, offset: Offset) -> bool {
-        let old_offset = self.offset();
-        let is_already_sealed = self.is_sealed();
-
-        let new_offset = std::cmp::max(self.offset(), offset);
-        let new_sealed = self.is_sealed() || sealed;
-        if new_sealed != is_already_sealed || new_offset > old_offset {
-            *self = TailState::new(new_sealed, new_offset);
-            true
-        } else {
-            false
-        }
-    }
-
-    /// Applies a seal on the tail state without changing the tail offset
-    /// Returns true if the state was updated
-    pub fn seal(&mut self) -> bool {
-        if self.is_sealed() {
-            false
-        } else {
-            *self = TailState::new(true, self.offset());
-            true
-        }
-    }
-}
-
-impl<Offset: SequenceNumber> TailState<Offset> {
-    pub fn map<F, T>(self, f: F) -> TailState<T>
-    where
-        F: FnOnce(Offset) -> T,
-    {
-        match self {
-            TailState::Open(offset) => TailState::Open(f(offset)),
-            TailState::Sealed(offset) => TailState::Sealed(f(offset)),
-        }
-    }
-
-    #[inline(always)]
-    pub fn is_sealed(&self) -> bool {
-        matches!(self, TailState::Sealed(_))
-    }
-
-    #[inline(always)]
-    pub fn offset(&self) -> Offset {
-        match self {
-            TailState::Open(offset) | TailState::Sealed(offset) => *offset,
-        }
-    }
-}
-
 /// A future that resolves to the Lsn of the last Lsn in a committed batch.
 ///
 /// Note: dropping this future doesn't cancel or stop the underlying enqueued append.
